% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cells.r
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Aggregate cells}
\usage{
aggregate_cells(X, Y = NA, dX, dY = NA, form = NA, slide = NULL,
  locs = NULL)
}
\arguments{
\item{X}{(required) x-dimension of the grid or a vector of length 2 giving 
the dimensions of the grid \code{(x, y)}}

\item{Y}{y-dimension of the grid}

\item{dX}{(required) number of rows to aggregate or a vector of length 2 
giving the dimensions to aggregate}

\item{dY}{number of columns to aggregate}

\item{form}{character string indicating how cells should be aggregated 
(see details).}

\item{slide}{if \code{form = 'window'}, a vector of length 1 or 2 indicating
the number of cells to slide the window in the x and y directions.
Default is one cell in each direction.}

\item{locs}{if \code{form = 'origin'}, a matrix of cell locations on 
which aggregations should be centered. Default is the center cell.}
}
\value{
a list of matrices
}
\description{
Groups cells of a grid together and returns a list
of cell locations.
}
\details{
This function is used for summarizing simulation output across different 
spatial grains. The function returns a list of matrices, where each matrix
specifies the locations of cells to be aggregated. This list can be passed
as the \code{locs} parameter in summary functions such as 
\code{\link{summarize_sim_N}} and the functions therein.
The function implements three methods for aggregating cells:
	\describe{
	\item{'partition'}{The grid is partitioned so that groups with 
		dimensions \code{(dX,dY)} are non-overlapping, starting from the  
		first row, first column. If the number of rows and columns in 
		the grid is not divisible by \code{dX} or \code{dY} then the 
		last aggregation in each row or column will contain fewer columns 
			or rows of cells than specified.}
	\item{'window'}{A window with dimensions \code{(dX, dY)} is slid
		repeatedly across the grid according the the number of steps given  
		in \code{slide}. If \code{slide} is less than \code{(dX, dY)}
		this results in overlapping groups of cells.}
	\item{'origin'}{Groups of cells with dimensions \code{(dX,dY)}   
		are aggregatedaround a focal cell or set of 
		cells given in \code{locs}.}
}
}
