% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_hab_type.r
\name{aggregate_hab_type}
\alias{aggregate_hab_type}
\title{Aggregate habitat types}
\usage{
aggregate_hab_type(land, locs)
}
\arguments{
\item{land}{(required) a raster landscape (as returned by 
\code{\link{make_landscape}}.}

\item{locs}{(required) either a two-column matrix of cell locations defining
a single spatial unit or a list of matrices designating
multiple spatial units}
}
\value{
a vector of habitat types corresponding to the spatial units
}
\description{
Determine the dominant habitat type in a set of spatial units
}
\details{
The function determined which habitat type is present in the majority of
cells in each spatial unit specified in the list \code{locs}.
}

