% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abun_profile.r
\name{calc_abun_profile}
\alias{calc_abun_profile}
\title{Abundance profiles}
\usage{
calc_abun_profile(locs, t_window, sim, N_S)
}
\arguments{
\item{locs}{(required) either a two-column matrix of cell locations where 
species' abundances should be summed or a list of matrices designating
multiple spatial units where abundance should be summed (see details)}

\item{t_window}{(required) either a list containing \code{start}
and \code{stop} specifying that all collected timepoints in that interval
should be considered or an explicit vector of timepoints}

\item{sim}{(required) array of simulation results from a single 
simulation run (as returned by \code{\link{run_sim}})}

\item{N_S}{(required) number of species in the simulation}
}
\value{
an array of species' abundances with dimensions:
	\code{[timepoint, species, spatial unit]}
}
\description{
Calculates species' abundances through time for a simulation.
}
\details{
The function calculates species' abundances over a window of time 
given in \code{t_window}. This can be either a continuous set of timepoints
(designated by passing a list with elements \code{start} and \code{stop})
or an explicit set of timepoints (passed as a vector of integers). 
Timepoints correspond to actual timesteps in the simulation, even if
not all timesteps are saved in \code{sim}. Specifying \code{start=1}
will return the first timestep, not the first step saved in \code{sim}.
If the simulation did not save the the first timestep then the function
will exit with an error message.

Species abundances are summed over the set of cell locations given in 
\code{locs}. If \code{locs} is a list (as generated by 
\code{\link{aggregate_cells}}) then each element of the list specifies a 
matrix of cells (a 'spatial unit') in which abundance should be summed. 
For example, to calculate species' abundances in each cell separately, 
pass a list where each element is the coordinates of a single cell. 
The function returns an array of species' abundances with dimensions:
\code{[timepoint, species, spatial unit]}. It will include a column labeled
\code{'0'} for the number of empty spaces.
}
