% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CT.r
\name{calc_rich_CT}
\alias{calc_abun_CT}
\alias{calc_rich_CT}
\title{Calculate core/transient richness and abundance}
\usage{
calc_rich_CT(abuns, occupancy, breaks, agg_times = NULL)

calc_abun_CT(abuns, occupancy, breaks, agg_times = NULL)
}
\arguments{
\item{abuns}{(required) array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}})}

\item{occupancy}{(required) matrix of species temporal occupancies at
a set of locations (as returned by \code{\link{calc_occupancy}})}

\item{breaks}{(required) a vector of numbers on \code{[0,1]} (in order) 
specifying boundaries of occupancy categories}

\item{agg_times}{either a single number specifing the number of timepoints 
that should be aggregated before calculating occupancy or a list of
vectors defining exactly which timepoints should be aggregated (see details).
 Defaults to no aggregation.}
}
\value{
an array with dimensions \code{[temporal units, occupancy 
	categories, spatial units]}
}
\description{
Calculates richness or abundance of core and transient species through 
time for a simulation
}
\details{
These functions calculate species richness and abundance of
species in different categories of temporal occupancy.
The spatial and temporal scope is determined by an array of
species abundance profiles (\code{abuns}, as returned by
\code{\link{calc_abun_profile}} while the matrix of temporal occupancies 
used to categorize species are given in 
\code{occupancy} (see \code{\link{calc_occupancy}}). 
Species are assigned to one of two or more categories defined by
\code{breaks}, which is an ordered list of numbers in 
\code{[0,1]} defining occupancy category boundaries. Occupancy category 
intervals include their upper bound, but not their lower bound.
Thus, species with occupancy = \code{0} are not counted.
Specifying a large number of breaks may be useful for generating
occupancy distributions.
Timepoints may be aggregated prior to calculations using 
\code{agg_times}. This can be useful for evaluating the effects 
of sampling at different temporal scales on observed richness
and abundance. This aggregation is relative to the timepoints
stored in \code{abuns} and not to the actual 
timesteps in the simulation. For example, specifying 
\code{agg_times=10} on abundance profiles containing
every other timestep will aggregate abundances from \code{T=1:20,
T=21:40,...}.
}
\section{Functions}{
\itemize{
\item \code{calc_rich_CT}: Calculate core-transient species richness

\item \code{calc_abun_CT}: Calculate core-transient abundance
}}

