% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_occupancy.r
\name{calc_occupancy}
\alias{calc_occupancy}
\title{Temporal occupancy}
\usage{
calc_occupancy(locs = NULL, t_window = NULL, sim = NULL,
  N_S = NULL, abuns = NULL, agg_times = NULL, which_species = NULL,
  do_freq = F)
}
\arguments{
\item{locs}{(required) either a two-column matrix of cell locations where 
species' abundances should be summed or a list of matrices designating
multiple spatial units where abundance should be summed (see details)}

\item{t_window}{(required) either a list containing \code{start}
and \code{stop} specifying that all collected timepoints in that interval
should be considered or an explicit vector of timepoints}

\item{sim}{(required) array of simulation results from a single 
simulation run (as returned by \code{\link{run_sim}})}

\item{N_S}{(required) number of species in the simulation}

\item{abuns}{array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}})}

\item{agg_times}{either a single number specifing the number of timepoints 
that should be aggregated before calculating occupancy or a list of
vectors defining exactly which timepoints should be aggregated (see details).
 Defaults to no aggregation.}

\item{which_species}{a vector of integers indicating which species should be
considered. Defaults to all species.}

\item{do_freq}{a logical indicating whether to return actual frequencies 
rather than occupancy (which is scaled by number of timepoints). 
Defaults to FALSE.}
}
\value{
a matrix with dimensions \code{[spatial units, species]}
}
\description{
Calculate species' temporal occupancy in a simulation.
}
\details{
The function calculates temporal occupancy for each species:
the fraction of timepoints a species is present in a location
during a given time window. Users can either supply an 
array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}}) or a set of spatial units
(\code{locs}), a time window (\code{t_window}), simulation
results (\code{sim}) and the number of species (\code{N_S}), 
from which abundance profiles can be calculated. These parameters
are explained in more detail in the documentation of 
\code{\link{calc_abun_profile}}. For each spatial unit in
the abundance profiles, the function calculates the number of
timepoints in which each species was present and if 
\code{do_freq = FALSE}, divides by the total number of timepoints. 
If \code{do_freq = TRUE} the actual number of timepoints is returned. 
Timepoints may be aggregated prior to this calculation using 
\code{agg_times}. This can be useful for evaluating the effects 
of sampling at different temporal scales on observed occupancy.
This aggregation is relative to the timepoints
stored in the abundance profiles array and not to the actual 
timesteps in the simulation. For example, specifying 
\code{agg_times=10} on abundance profiles containing
every other timestep will aggregate abundances from \code{T=1:20,
T=21:40,...}.
}
