% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_species_turnover.r
\name{calc_species_turnover}
\alias{calc_species_turnover}
\title{Species turnover}
\usage{
calc_species_turnover(turnover, locs, t_window, agg_times = NULL,
  which_species = NULL)
}
\arguments{
\item{turnover}{(required) an array of species rates of turnover, as
    returned by \code{run_sim} when \code{calc_rates=TRUE}. Should have
dimensions \code{[time, row, column, species, gain/loss rate]}.}

\item{locs}{(required) either a two-column matrix of cell locations where 
species' rates should be summed or a list of matrices designating
multiple spatial units where rates should be summed (see details)}

\item{t_window}{(required) either a list containing \code{start}
and \code{stop} specifying that all collected timepoints in that interval
should be considered or an explicit vector of timepoints}

\item{agg_times}{either a single number specifing the number of timepoints 
that should be aggregated before calculating rates or a list of
vectors defining exactly which timepoints should be aggregated (see details).
Defaults to no aggregation.}

\item{which_species}{a vector of integers indicating which species should be
grouped and averaged. Defaults to each species analyzed separately.}
}
\value{
a array of rates of species gains and losses in each location at 
	each timepoint (dimensions \code{[temporal unit, spatial unit, 
	gain or loss, species group]}).
}
\description{
Calculate rates of species' gains and losses
}
\details{
DOCUMENTATION NOT DONE
The function calculates the number of individuals in a group of 
species that are gained or lost per timestep. The function takes 
a turnover matrix, returned by \code{\link{run_sim}} when
code{calc_rates=TRUE} and then calculates the number of individuals
of each species that are gained or lost per timestep, summing across
cells in each spatial unit in \code{locs}. A parameter \code{agg_times}
can be used to aggregate timepoints prior to calculating rates. 
This aggregation is relative to the timepoints stored in the turnover
array and not to the actual timesteps in the simulation. 
For example, specifying \code{agg_times=10} on a turnover matrix containing
every other timestep will aggregate individuals from \code{T=1:20,
T=21:40,...} and then divide each by 20 timesteps to calculate the rates.
Once individual species rates have been calculated, the function sums
rates across species groups (\code{which_species}), given as a vector 
that can be interpreted as a factor identifying which species 
belong to which groups. If the user wishes to calculate
average rates within groups they will need to divide the results by
the number of species in each group. By default species are not grouped and 
rates for individual species are returned.
}

