% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_classify.r
\name{cross_classify}
\alias{cross_classify}
\title{Classify occupancy vs residency}
\usage{
cross_classify(occupancy, breaks, b_rates = NULL, habitats = NULL,
  classification = NULL, do_each = F, return = "counts")
}
\arguments{
\item{occupancy}{(required) matrix of species temporal occupancies at
a set of locations (as returned by \code{\link{calc_occupancy}})}

\item{breaks}{(required) either a numeric vector of length 1 or 2 giving 
occupancy breakpoints for transient vs. core status or a named list with 
specific intervals: \code{list(trans=c(), core=c())}}

\item{b_rates}{matrix of birth rates for each species in each habitat type.
Dimensions are \code{[species, habitat type]}.}

\item{habitats}{vector of habitat types (\code{'A'} or \code{'B'}) 
in each location}

\item{classification}{matrix classifying each species at each location as 
\code{'core'} or \code{'trans'}. Dimensions are 
\code{[locations, species]}.}

\item{do_each}{logical indicating whether one table should be returned 
summarizing across all locations (\code{FALSE}) or an array with one table 
for each spatial unit \code{TRUE}. Defaults to \code{FALSE}.}

\item{return}{character string indicating whether a table of \code{'counts'}
or lists of species IDs (\code{'ids'}) should be returned. 
Default is \code{'counts'}.}
}
\value{
table or array, depending on \code{do_each}
}
\description{
Cross-classify species based on their temporal occupancy
and true status as core residents or transients
}
\details{
This function classifies species as core or transient at a set of locations
based on temporal occupancy and also as as biologically core or transient
based on birth rates, \code{b_rates}, or a pre-defined 
\code{classification}. It then counts the number of species occuring 
in each of the four categories: biologically core - occupancy core, 
biologically core - occupancy transient, biologically transient - occupancy 
core, and biologically transient - occupancy transient. 
Occupancy-based core/transient categories are determined by breakpoints 
given in \code{breaks}. This can be either a named list with explicit
intervals or a vector of one or two numbers on the interval
\code{[0,1]}. In this case, transient species are those with occupancies 
less than or equal to \code{breaks[1]} while core species are those with 
occupancies greater than \code{max(breaks)}. The status of species as 
biologically core or transient is specified either by a matrix classifying  
each species as \code{'core'} or \code{'trans'} at each spatial location
(\code{classification}) or by supplying both \code{b_rates} and a vector of
the habitat types at each location (\code{habitats}). This should be
calculated for the same set of locations in \code{occupancy} using 
\code{\link{average_habitat}}.

The function can return either a count of the number of species in each category
or a list of the species that are in each category (specify with 
\code{return}). It can also return a separate table for each location if
\code{do_each=TRUE}.
}
