% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.r
\name{die}
\alias{die}
\title{Death}
\usage{
die(comm, m_rates)
}
\arguments{
\item{comm}{(required) vector (or list) of species present
(may contain zeros for empty spaces)}

\item{m_rates}{(required) vector of mortality rates for each species
in this cell (may be habitat specific)}
}
\value{
a vector of integers of the same lenth as \code{code}
	where 0s replace individuals that have died
}
\description{
Determines which individuals in a community die during one time step.
}
\details{
\code{comm} is a vector (or list) of integers indicating which species 
are present. Individuals in the community die with species-specific 
probabilities provided in \code{m_rates}. The order of \code{m_rates}  
designates which species each mortality rate applies to. 
For example, \code{m_rates[1]} is the mortality rate of species 1.
}
