% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_habitat.r
\name{get_habitat}
\alias{get_habitat}
\title{Determine habitat type}
\usage{
get_habitat(x)
}
\arguments{
\item{x}{(required) a number}
}
\value{
'A' or 'B'
}
\description{
Determines the habitat type that corresponds to a numeric value
(as stored in a landscape).
}
\note{
Does not work directly on vectors. Use \code{sapply(x, get_habitat)}.
}

