% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sptype.r
\name{get_sptype}
\alias{get_sptype}
\title{Determine species type}
\usage{
get_sptype(species)
}
\arguments{
\item{species}{(required) array of species vital rates}
}
\value{
a vector of characters indicating which habitat each species 
	prefers: 'A', 'B', or 'AB' (for a generalist).
}
\description{
Determines whether each species in a species pool is a specialist on
habitat A or B, or a generalist using birth rates.
}
\details{
User should supply an array \code{species} in which the third dimension
	has a column named 'b' that stores birth rates (such as returned by 
	\code{\link{make_species}}.)
}
\seealso{
\code{\link{make_species}}
}

