% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_landscape.r
\name{make_landscape}
\alias{make_landscape}
\title{Creation of new landscape}
\usage{
make_landscape(x = NULL, y = NA, mod = NULL, d = NA, prop = NA,
  draw_plot = F)
}
\arguments{
\item{x}{(required) number of cells in the x dimension, or a vector of the x and y
dimensions}

\item{y}{number of cells in the y dimension}

\item{mod}{variogram model used to generate spatial autocorrelation 
(see  \code{\link{vgm}}).
Defaults to Exponential model with partial sill = 1.}

\item{d}{range of the variogram model. Defines the distance at which cells
are correlated. Defaults to 1/3 of the smallest dimension.}

\item{prop}{proportion of cells assigned to have value = 1.}

\item{draw_plot}{logical indicating whether function should plot the
landscape}
}
\value{
a raster object
}
\description{
Creates a landscape as a grid with two habitat types represented by a raster
	 of 1 and -1.
}
\note{
Habitat heterogeneity increases as the ratio between \code{d} and grid 
	dimensions decreases. WARNING: long compute time on grids > 50 x 50.
}
\seealso{
\code{\link{vgm}} for how to specify variogram model
}

