% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sad.r
\name{make_sad}
\alias{make_sad}
\title{Abundance distribution generation}
\usage{
make_sad(N_S, distribution)
}
\arguments{
\item{N_S}{number of species}

\item{distribution}{named list of  parameters describing the functional 
form of the distribution (SAD).}
}
\value{
a vector of integers of length \code{N_S}
}
\description{
Samples a given distribution for a set number of species abundances.
Implements several commonly used distributions (see details).
}
\details{
The form of the distributition is controlled by a named list with
	 the following components:
\describe{
	\item{\code{type}}{(Required) indicates the form of the SAD.
		Can be: 'same', 'uniform', 'power', 'logseries', 'lognormal',
			'poisson'}
	\item{\code{maxN}}{maximum abundance, used to find SAD parameters
		(if not supplied as p1 or p2)}
	\item{\code{P_maxN}}{probability of getting an abundance greater than maxN}
	\item{\code{p1}}{parameter controling the shape of the SAD
		(only certain types)}
	\item{\code{p2}}{parameter controling the shape of the SAD
		(only certain types)}
}
}
\note{
Code is set up so that abundance may be correlated with another
	using Gaussian copulas, but this is currently not implemented. 
	To add a correlation, generate x as a multivariate random normal 
	variable with specified covariance matrix.
}

