% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_landscape.r
\name{populate_landscape}
\alias{populate_landscape}
\title{Create a metacommunity}
\usage{
populate_landscape(land, species, gsad = NULL, K, distribution = NA,
  p = NA, which_cells = NULL)
}
\arguments{
\item{land}{(required) matrix or raster of habitat types defining
the landscape (as generated by \code{\link{make_landscape}})}

\item{species}{(required) array of species vital rates (as generated by
\code{\link{make_species}})}

\item{gsad}{vector defining the global relative abundance of each species.
    Must be in the same order as \code{species}. Defaults to same abundance 
for all species.}

\item{K}{(required) a single number defining a carrying capacity for all cells
 or a matrix of carrying capacities for each landscape cell. Must match
dimension of \code{land}.}

\item{distribution}{character describing how individuals should be 
disributed across the landscape (see details). Defaults to \code{'same'}  
unless \code{which_cells} is specified.}

\item{p}{proportion of carrying capacity to be filled. Defaults to 1 = full.}

\item{which_cells}{two column matrix indicated the x and y coordinates of 
cells to receive individuals. May optionally contain a 3rd column 
specifying the number of individuals in each cell.}
}
\value{
a matrix of lists defining the metacommunity (see details).
}
\description{
Creates a community for each cell on a landscape by defining the number  
of individuals each cell can hold (carrying capacity) and
distributing individuals from a species pool into those communities
according to a global species abundance distribution (gsad).
}
\details{
The function returns a metacommunity on which a simulation can be conducted.
A metacommunity is a matrix of lists whose dimensions match the landscape
specified by \code{land}. Each list corresponds to a cell and contains a 
single vector of integers indicating which species are currently 
established in the cell.

Individuals are distributed across the landscape in two steps.
First, the function determines a carrying capacity for each cell using
the parameter \code{K}, which can be either a single number, for the same
number of individuals in each cell, or a matrix whose dimensions match the
landscape giving a different number of individuals in each cell.
The spatial distribution of individuals is then defined by the 
\code{distribution} parameter which can have the following values:
\describe{
	\item{\code{'same'}}{all cells are filled to the same proportion 
			(\code{p}) of their carrying capacity.}
	\item{\code{'uniform'}}{individuals are randomly placed in cells such
		such that the total carrying capacity of the landscape is filled
		to proportion \code{p}}
	\item{\code{'designated'}}{user specifies which cells should receive 
		individuals and (optionally) how many individuals each cell should
		receive using the \code{which_cells} parameters. If 
		is only specified with two columns (for the cell locations)
			then each cell is filled to proportion \code{p}.}
}
Species are chosen at random according to their relative abundance, 
as specified by \code{gsad}.
}

