% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.r
\name{reproduce}
\alias{reproduce}
\title{Generate propagules}
\usage{
reproduce(comm, b_rates)
}
\arguments{
\item{comm}{(required) vector (or list) of species present
(may contain zeros for empty spaces)}

\item{b_rates}{(required) vector of birth rates for each species}
}
\value{
a vector of integers, one for each propagule
}
\description{
Generates a pool of propagules produced by a community during one time step.
}
\details{
\code{comm} is a vector of integers indicating which species are present.
The order of \code{b_rates} designates which species each birth rate
applies to. For example, \code{b_rates[1]} is the birth rate of species 1.
}

