% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_P.r
\name{run_sim_P}
\alias{run_sim_P}
\title{Run multiple parameter sets}
\usage{
run_sim_P(ncores = 1, parm_dir = "./", results_dir = "./Results/",
  sim_dir = NULL, report = 0, restart = F)
}
\arguments{
\item{ncores}{number of cores to use for parallel simulation.
Defaults to 1 for serial simulation.}

\item{parm_dir}{directory where parameter files are located.
Defaults to the current directory.}

\item{results_dir}{directory where results should be saved.
Defaults to 'Results' in the current directory.}

\item{sim_dir}{directory where \code{CTSim} is installed, 
if not on the default search path}

\item{report}{timestep interval at which to report simulation status
to stdout. Defaults to 0 for no reporting.}

\item{restart}{logical indicating whether this a restart of a previously
existing set of runs. Defaults to \code{FALSE}}
}
\value{
nothing
}
\description{
Run simulations on multiple sets of parameters
}
\details{
This function runs simulations on parameter files stored in the directory
given in \code{parm_dir}. Parameter filenames must start with 'p_'.
Each parameter file should have a unique \code{simID} defined within it.
Results are saved in the directory \code{results_dir} and are not returned
by the function.
}
\seealso{
\code{\link{run_simulation}} and \code{\link{restart_simulation}}
for command line execution
}

