% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\arguments{
\item{ncores}{number of cores to use for parallel simulation.
Defaults to 1 for serial simulation.}

\item{parm_dir}{directory where parameter files are located.
Defaults to the current directory.}

\item{results_dir}{directory where results should be saved.
Defaults to 'Results' in the current directory.}

\item{sim_dir}{directory where \code{CTSim} is installed, 
if not on the default search path}

\item{report}{timestep interval at which to report simulation status
to stdout. Defaults to 0 for no reporting.}
}
\description{
Script used to run simulations on a set of parameter files
from the command line using \code{R CMD BATCH}.
}
\details{
This script is found in the exec/ directory where the package is 
installed. To run, move script to working directory and use:

\code{R CMD BATCH '--args ncores parm_dir results_dir sim_dir report'  
	run_simulation.R outfile.Rout}
}
\seealso{
\code{\link{run_sim_P}} for a description of parameters 
that can be passed through args
}
