% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sim.r
\name{sample_sim}
\alias{sample_sim}
\title{Sample simulation with imperfect detection}
\usage{
sample_sim(abuns, probs = NULL, return = "abundance")
}
\arguments{
\item{abuns}{(required) array of abundance profiles (as returned by 
\code{\link{calc_abun_profile}})}

\item{probs}{probability of observing an individual. Can be either a 
single detection probability for all species, or a vector with different 
probabilies for each species. Defaults to 1.}

\item{return}{a character string indicating whether species abundances 
(\code{'abundance'}) or presence (\code{'presence'}) should be returned. 
Default is abundance.}
}
\value{
an array with the same dimensions as \code{abuns}
}
\description{
Stochastically samples a simulation using a fixed detection probability.
}
\details{
The function samples the abundances of species at 
a set of locations and timepoints (given in \code{abuns}) using 
detection probabilities in \code{probs}. These are the probabilities
that an individual of a given species will be observed and can be set to
be the same for all species by only providing a single number. It returns
an array with the same dimensions as \code{abun} containing either the
observed abundances or presence/absence (depending on \code{return}).
}

