% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sim.r
\name{summarize_sim}
\alias{summarize_sim}
\title{Summarize a simulation}
\usage{
summarize_sim(sim, breaks, locs, t_window, species = NULL, land = NULL,
  gsad = NULL, agg_times = NULL, P_obs = list(1), sum_parms = NULL,
  sum_turn = TRUE)
}
\arguments{
\item{sim}{(required) either a filename for a simulation run (as saved by
\code{\link{run_sim_N}} or an array of simulation results (as returned by 
\code{\link{run_sim}}. If an array, must specify \code{species}, 
\code{land} and \code{gsad}.}

\item{breaks}{(required) a vector of numbers on \code{[0,1]} (in order) 
specifying boundaries of occupancy categories}

\item{locs}{(required) either a two-column matrix of cell locations where 
species' abundances should be summed or a list of matrices designating
multiple spatial units where abundance should be summed (see 
\code{\link{calc_abun_profile}})}

\item{t_window}{(required) either a list containing \code{start}
and \code{stop} specifying that all collected timepoints in that interval
should be considered or an explicit vector of timepoints}

\item{species}{array of species vital rates used in the simualation
(as generated by \code{\link{make_species}})}

\item{land}{matrix or raster of habitat types defining
the landscape used in the simulation
(as generated by \code{\link{make_landscape}})}

\item{gsad}{vector defining the global relative abundance of each species
used in the simulation. Must be in the same order as \code{species}.}

\item{agg_times}{either a single number specifing the number of timepoints 
that should be aggregated before calculating occupancy or a list of
vectors defining exactly which timepoints should be aggregated (see details).
 Defaults to no aggregation.}

\item{P_obs}{vector of detection probabilities (see details and 
\code{\link{sample_sim}})}

\item{sum_parms}{list of parameters controlling how the simulation is 
    summarized across spatial and temporal units. May contain:
    \describe{
        \item{agg_times}{specifies how time points should be aggregated 
        before calculations. See \code{\link{calc_rich_CT}} for details.}
    \item{time_sum}{character vector indicating which timepoint
        should be used in summary statistics (see details)}
    \item{quants}{vector of quantiles for summarizing across spatial units
        (see details)}
    \item{hab}{character string indicating whether summary should be performed
        across all habitat types (default) or for just one (indicated by 
        \code{'A'} or \code{'B'}.}
}}

\item{sum_turn}{logical indicating whether turnover rates should be
summarized. Defaults to \code{TRUE}.}
}
\value{
a list of arrays defined as follows:
	\describe{
		\item{bio}{richness and abundance of biologically defined species,
			has dimensions: \code{[richness or abundance, spatial 
			summary statistic, core or transient, detection probability]}}
		\item{occ}{richness and abundance in temporal occupancy categories,
			has dimensions: \code{[richness or abundance, spatial 
			summary statistic, occupancy category, detection probability]}}
		\item{xclass}{core-transient cross-classification,
			has dimensions: \code{[spatial summary statistic,
			 cross-classification category, detection probability]}}
		\item{abun}{species' relative abundances, has dimensions:
			\code{[species, detection probability]}}
	}
	If \code{sum_parms$time_sum = 'none'} then each of these arrays will
	have an additional dimension corresponding to timepoints.
}
\description{
Summarizes the output of a single simulation run.
}
\details{
The function calculates summary statistics of four main quantities and 
returns them in a named list:
\describe{
\item{bio}{richness and abundance of biologically core and transient 
	species (based on birth rates)}
\item{occ}{richness and abundance of species in classes based on
	their temporal occupancy}
\item{xclass}{number of species in each of four categories: biologically
	core - occupancy core, biologically core - occupancy transient,
	biologically transient - occupancy core, and biologically transient -
	occupancy transient (see\code{\link{cross_classify}})}
\item{abun}{relative rank abundance of each species
	in the metacommunity}
}

The function proceeds as follows:
\enumerate{
\item Actual species' abundances are calculated for each spatial unit 
	in \code{locs} and temporal unit in \code{t_window}. See documentation
	of \code{\link{calc_abun_profile}} for more information on 
	these parameters.
\item Observed species' abundance profiles are calculated by sampling
	actual abundance profiles using the detection probabilities 
	provided in \code{P_obs}.
\item Observed abundance profiles are used to calculate species 
	temporal occupancy in each of the spatial units, after aggregating 
	timepoints according to \code{agg_times} (see 
	\code{\link{calc_occupancy}}).
\item Richness and abundance of species in each of the temporal occupancy
	categories defined by \code{breaks} are calculated in each spatial 
	unit for each timepoint. Timepoints may
	be aggregated prior to calculations using the \code{agg_times}
	argument in \code{sum_parms}. See (see \code{\link{calc_rich_CT}}) for 
	details.
\item For each spatial unit, habitat values are averaged across the
	cells that comprise it (see \code{\link{average_habitat}} and this 
	average habitat type is used to determine which species are 
	biologically core and transient (based on whether they have positive
	birth rates in that habitat). Richness and abundance of biologically
	core and transient species is then calculated for each spatial unit
	at each timepoint. As in the previous step, timepoints may be 
	aggregated prior to calculations using the \code{agg_times} 
	argument in \code{sum_parms}.
\item For each spatial unit, species are classified as core or transient 
	based on whether they fall in the first or last occupancy interval defined
	by \code{breaks}. Then, the number of species classified as biologically
	core - occupancy core, biologically core - occupancy transient,
	biologically transient - occupancy core, and biologically transient -
	occupancy transient are counted for each spatial unit.
\item The element \code{time_sum} in \code{sum_parms} defines which 
	timepoints are used for summarizing richness and abundance. If 
	\code{time_sum = 'none'} then all timepoints are summarized 
	individually. Other options are \code{'mean'} for averaging across all
	timepoints in \code{t_window} or \code{'last'} for only using the
	last timepoint in \code{t_window}.
\item Summary statistics are calculated across all spatial units.
	Minimally, this is the mean and variance, but may also include any
	quantiles specified in the \code{quants} argument of \code{sum_parms}.
\item The relative abundance of each species is calculated across the entire
	metacommunity for each set of observed abundances (from step 2).
}
}
