% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_simulation_cross_time.R
\name{summarize_simulation_cross_time}
\alias{summarize_simulation_cross_time}
\title{Summarize simulations through time}
\arguments{
\item{run_dir}{directory where simulation runs are located.
Defaults to the current directory.}

\item{parm_dir}{directory where parameter files are located.
Defaults to the current directory.}

\item{results_dir}{directory where results should be saved.
Defaults to 'Summaries' in the current directory.}

\item{sim_dir}{directory where \code{CTSim} is installed, 
if not on the default search path}
}
\description{
Script used to summarize simulations on a set of parameter files
through time from the command line using \code{R CMD BATCH}.
}
\details{
This script is found in the exec/ directory where the package is 
installed. To run, move script to working directory and use:

\code{R CMD BATCH '--args ncores run_dir parm_dir results_dir sim_dir'  
	summarize_simulation_cross_time.R outfile.Rout}
}
\seealso{
\code{\link{summarize_sim_P}} for a description of parameters 
that can be passed through args
}

