% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.r
\name{run_sim}
\alias{run_sim}
\title{Run simulation}
\usage{
run_sim(steps, metacomm, land, species, gsad, d_kernel = NULL,
  v_kernel = NULL, imm_rate = NA, save_steps = NULL, report = 0,
  ID = NA)
}
\arguments{
\item{steps}{(required) number of timesteps to run the simulation}

\item{metacomm}{(required) matrix of lists defining metacommunity
(as generated by \code{\link{populate_landscape}} or returned by
this function).}

\item{land}{(required) matrix or raster of habitat types defining
the landscape (as generated by \code{\link{make_landscape}})}

\item{species}{(required) array of species vital rates (as generated by
\code{\link{make_species}})}

\item{gsad}{(required) vector defining the global relative abundance of 
each species. Must be in the same order as \code{species}. 
Defaults to same abundance for all species.}

\item{d_kernel}{list defining the dispersal kernel of new propagules (see 
\code{\link{get_dispersal_vec}} for options). Passed to the \code{form} 
parameter in \code{\link{disperse}}. Defaults to Gaussian.}

\item{v_kernel}{list defining the dispersal kernel of previously established
individuals (see \code{\link{get_dispersal_vec}} for options). Passed to 
the \code{form} parameter in \code{\link{disperse}}. Defaults to Gaussian.}

\item{imm_rate}{immigration rate. Passed to the parameter \code{m} in
\code{\link{establish}}. Defaults to 0.}

\item{save_steps}{vector of timesteps at which to save the simulation. 
Defaults to all timesteps, including the initial metacommunity.}

\item{report}{integer giving the interval at which the function should 
report its progress to \code{stdout}. Defaults to 0 for no reporting.}

\item{ID}{character string identifying this simulation run when reporting progress}
}
\value{
a 3-dimensional array of lists defining the metacommunity at each 
	timepoint in \code{save_steps}. Time is in the 3rd dimension.
}
\description{
Runs the simulation on a metacommunity for a fixed number of timesteps.
}
\details{
Given a metacommunity (\code{metacomm}, see 
\code{\link{populate_landscape}}), landscape (\code{land}, 
see \code{\link{make_landscape}}), species pool (\code{species}, see
\code{\link{make_species}}), and global species abundance distribution
(\code{gsad}), this function calls \code{\link{run_timestep}}
to progress a metacommunity through a fixed number of timesteps (given by 
\code{steps}). The user may specify which timesteps should be returned using the
\code{save_steps} parameter. Including \code{0} in \code{save_steps} will  
save the initial metacommunity.
}
\seealso{
\code{\link{run_timestep}} for how the simulation progresses 
	through a single timestep. \cr
\code{\link{run_sim_N}} for running multiple independent simulations on 
	the same set of parameters.
}

